--- solr/ui/src/commonMain/kotlin/org/apache/solr/ui/components/configsets/data/ListConfigSets.kt
@@ -1,13 +1,0 @@
-package org.apache.solr.ui.components.configsets.data
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-/**
- * List of config sets.
- */
-@Serializable
-data class ListConfigSets(
-    @SerialName("configSets")
-    val names: List<String> = emptyList(),
-)
