+++ solr/ui/src/commonMain/kotlin/org/apache/solr/ui/views/configsets/OverviewContent.kt
@@ -14,6 +14,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.solr.ui.components.configsets.overview
+package org.apache.solr.ui.views.configsets
 
-interface OverviewComponent
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import org.apache.solr.ui.components.configsets.overview.OverviewComponent
+
+@Composable
+fun OverviewContent(component: OverviewComponent, modifier: Modifier = Modifier) {
+    Text("Overview section")
+}
